Integrated blockchain into select merchant page
