--- src/ethereum/components/SelectMerchant.js
@@ -1,45 +1,0 @@
-import React, { Component } from 'react'
-import web3 from '../web3'
-import SmartGift from '../smartgift'
-
-class SelectMerchant extends Component {
-	state = {
-		errorMessage: ''
-	}
-
-	onSubmit = async (event) => {
-		event.preventDefault()
-
-		try {
-			const targetGift = SmartGift('0xd16038d71B68E149B9441dcEEf6C9c8b339701a6') // address of the Gift you're working on
-			const accounts = await web3.eth.getAccounts()
-			const selectionResult = await targetGift.methods
-				.recipientPicksMerchant('0xBae217221CbFE934d8e190c8f0A836Cd44a7ed07') // address of merchant that gets selected
-				.send({
-					from: accounts[0],
-					gas: 1000000
-				})
-			if (selectionResult.status === "0x0" || !selectionResult.status) {
-				console.log("Transaction Failed!!!")
-			} else {
-				console.log(`You've selected and paid the Merchant!`)
-				alert('Success!')
-			}
-		} catch (err) {
-			this.setState({ errorMessage: err.message })
-		}
-	}
-
-	render() {
-		return (
-			<div>
-				<form onSubmit={this.onSubmit}>
-					<button>Select Merchant</button>
-				</form>
-				<h1>{this.state.errorMessage}</h1>
-			</div>
-		)
-	}
-}
-
-export default SelectMerchant
